added sources for nanyc c++/bootstrap, a c++ compiler for the Nany Language
This is the first draft of the Nany compiler. It is written in C++ and
does not intend to implement every features of the Nany Programming
Language. It will be used for bootstraping the real nany compiler
(written in Nany).

For compiling it on your platofrm, you can refer to `readme.md`