added pass to update the operand type for opcode 'stackalloc'
The type provided by the opcode 'stackalloc' (after code instanciation)
is currently unused (but should be in the future by other backends,
like LLVM) and was always set to 'any'. To preserve the datatype, a
simple pass has been added after the instanciation of the atom
(class, func) to update this value accordingly.