added automatic integration of nsl:std.core
A minimal part of the NSL (Nany Standard Library) is required to compile
even a simple program with Nany. Some mecanism of the language are
actually implemented by the library and not the compiler itself (in a
compiler, the less the better).

Nany is compiler. But it intends to be embedded in C/C++ applications as
well, and in this case, having extra source files is a nuisance. For
that reason, nsl:std.core is now directly available from the C++ source
code in `libnany` and is automatically added to the default target when
a context is created.