do not always release the 2nd param with shortcircuit evaluation
The label (used for jumping at the end of the evaluation when the first
parameter is enough to determine the outcome of the condition) must be
located *after* the release of any variables from the 2nd parameter but
before any release for the first one.

The solution is to simply stack scopes for each parameter and to insert
a label just before the final 'end-of-scope'.