added pseudo stack size for units
The generated IR code relies on local variables (starting with '%': %1,
%2, $3...), for both static (typing) and dynamic (the program itself) code
However, to reduce memory allocation during the compilation, and mainly
to make sure that the generated code matches the initial expectations,
the maximum number of local variables must be declared first.

However, the implementation is currently strongly context-based (e.g.
always attached to a class or a function) and obviously requires some
modifications to allow global typedefs or variables.