suppress spurious err messages when a dtor can't be called
It may happen that the destructor of an object can't be called (more
precisely, can't be instanciated due to some errors). Those errors are
usually detected when destroying an object (via 'unref' opcode).

Without this patch, 2 error messages can be emitted, one about the dtor
itself, and another one about the opcode itself, which is completely
useless for a normal user.