added opcode 'ensureresolved'
The typing is mostly done on the fly. However it is not always possible,
like for function resolutions (where the parameters are required to
select the most appropriate version). The partial type resolution can
currently lead to the generation of invalid IR code. In this case:

	func foo -> 42;

	public func main
	{
		ref x = foo;
	}

`foo` will be considered as a mere variable even if properly identifier
as a function. The type must be modified 'a posteriori'.

The situation can be worse if another `foo` function would be available
(overload resolution).

The role of the opcode `ensureresolved` allows to finish the work and to
generate error messages accordingly.

side note: The same will happen for classes with generic type
parameters, since they will behave like normal parameters.