fixed a particular case with a combination 'any'/'void' return values
Fixes the following scenario:

	func foo {} // returns 'void'
	func bar -> foo();  // implicit return type 'any', translated into 'void'

So obviously, the transformation from 'any' to 'void' should be
accepted.

Assignation between 'void' values must be accepted as well. It is
especially convenient for generic programming, but it is also a usual
scenario for IR generated from the AST:

	// ...
	%14 = call foo
	// ...
	%15 follows -> %1
	%15 inherits qualifiers from %1
	%15 = assign %14
	return %15

As seen in the original source code, %1 (the return type) is 'void', as
%15 should be.