fixed invalid label ids for conditions
Label ids are expected to be strictly ordered (for performance reasons
when executing the program - this limitation should be removed when the
IR code for programs will bne revamped). As a consequence, the lvid for
the label must be created when generating the new label, even if used by
the previous opcodes.

Fixes #31