fixed memory.realloc intrinsic
Since opcdes can only contain 3 operands (the first one is the opcode
value + 3 operands to keep memory alignment), the `!!memory.reallocate`
intrinsic was resetting the first parameter input to set the value of
the new pointer. However, the first parameter was never reachable since
the IR produced from the AST generated a new local variable for each
parameter (required when a parameter is an real expression).
The behavior is still the same (updating the input register) but now the
first parameter is actually the returned value.