views: do no longer call 'empty()' from the view
When iterating through the view, the method `empty()` was called first
to unnecessary code execution. Additionnally, this call was supposed
to handle the 'else' case of the for-loop (for <something> do {} else {})

But this call is redundant with `findFirst()`, which may returns false
(if the predicate filters everything, even if the container is not
empty).