fixed strict parameter types resolution
Parameter types may be strictly typed or not. When it is not, the type is
deduced on the fly while instanciating the code. Strict typing is
obviously used for func overloading deduction:

    func foo(cref a) { print("some value: \(a)"); }
    func foo(cref a: string) { print("specific overload: \(a)"); }
    foo("dummy");

The previous implementation was working for simple type like that but
was suffering from severe limitations, since it was not really using the
opcodes to resolve the type.

The following example did not compile previously:

    func foo(cref a) {}
    func foo(cref a: std.Ascii) {}
    func main { foo('a'); }

And an invalid atom name was often provided for error reporting ('any'
instead of the given type)