stringrefs: use yuni::String instead of custom implementation
The initial intent was to reduce the memory overhead induced by common
string implementation. Since those strings won't be modified, 4 or 8 bytes
per string can be saved in getting rid of the internal buffer capacity.
This optimization is certainly negligible in front of SSO
(small string optimization) now widely (or about to be) implemented.